requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
APScheduler==3.9.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
black==22.6.0
braceexpand==0.1.7
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
discord-ext-bot==1.0.1
discord.py==1.7.3
fastapi==0.79.0
frozenlist==1.3.1
h11==0.13.0
idna==3.3
import-expression==1.1.4
importlib-metadata==4.12.0
influxdb-client==1.31.0
jishaku==2.5.1
line-profiler==3.5.1
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pydantic==1.9.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2
pytz-deprecation-shim==0.1.0.post0
rel==0.4.7
Rx==3.2.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
uvicorn==0.18.2
websocket-client==1.3.3
websockets==10.3
yarl==1.8.1
zipp==3.8.1
